refactor(database ): layout for database pages
