+++ apps/Conduit-UI/src/components/database/custom-endpoints/EndpointsList.tsx
@@ -1,17 +1,11 @@
 import React, { FC, useEffect, useRef } from 'react';
-import { FixedSizeList as List, ListChildComponentProps } from 'react-window';
-import InfiniteLoader from 'react-window-infinite-loader';
-import AutoSizer from 'react-virtualized-auto-sizer';
 import { debounce } from 'lodash';
 import { useAppDispatch, useAppSelector } from '../../../redux/store';
-import { Box, ListItemButton, ListItemIcon, ListItemText, Paper, Typography } from '@mui/material';
-import { OperationsEnum } from '../../../models/OperationsEnum';
-import { getOperation } from '../../../utils/getOperation';
-import { Skeleton } from '@mui/material';
 import {
   asyncAddCustomEndpoints,
   asyncSetCustomEndpoints,
 } from '../../../redux/slices/databaseSlice';
+import InfiniteScrollList from './InfiniteScrollList';
 
 const timeoutAmount = 750;
 
@@ -70,105 +64,16 @@ const EndpointsList: FC<Props> = ({ handleListItemSelect, search, operation, sel
     debouncedGetApiItems(endpoints.length, limit);
   };
 
-  const EndpointRow = ({ index, style }: ListChildComponentProps) => {
-    const endpoint = endpoints[index];
-
-    const getBadgeColor = (endpointForBadge: any) => {
-      switch (endpointForBadge.operation) {
-        case OperationsEnum.POST:
-          return '#49cc90';
-        case OperationsEnum.PUT:
-          return '#fca130';
-        case OperationsEnum.DELETE:
-          return '#f93e3e';
-        case OperationsEnum.GET:
-          return '#61affe';
-        case OperationsEnum.PATCH:
-          return '#50e3c2';
-      }
-    };
-
-    return (
-      <div style={style}>
-        {!endpoint ? (
-          <Typography variant="h2">
-            <Skeleton />
-          </Typography>
-        ) : (
-          <ListItemButton
-            sx={{
-              '&.MuiListItemButton-root': {
-                '&.Mui-selected': {
-                  background: 'secondary',
-                },
-                borderRadius: '10px',
-              },
-            }}
-            key={`endpoint-${endpoint._id}`}
-            onClick={() => handleListItemSelect(endpoint)}
-            selected={selectedEndpoint?._id === endpoint?._id}>
-            <ListItemIcon>
-              <Paper
-                elevation={12}
-                sx={{
-                  color: 'white',
-                  padding: 0.2,
-                  width: '65px',
-                  textAlign: 'center',
-                  marginRight: 1,
-                  backgroundColor: getBadgeColor(endpoint),
-                }}>
-                {getOperation(endpoint)}
-              </Paper>
-            </ListItemIcon>
-            <ListItemText
-              primary={endpoint.name}
-              primaryTypographyProps={{
-                style: { overflow: 'hidden', whiteSpace: 'nowrap', textOverflow: 'ellipsis' },
-              }}
-            />
-          </ListItemButton>
-        )}
-      </div>
-    );
-  };
-
   return (
-    <AutoSizer>
-      {({ height, width }) => {
-        if (!count) {
-          return (
-            <Box width={width}>
-              <Typography sx={{ textAlign: 'center', marginTop: '100px' }}>
-                No endpoints available
-              </Typography>
-            </Box>
-          );
-        }
-        return (
-          <InfiniteLoader
-            ref={infiniteLoaderRef}
-            isItemLoaded={isItemLoaded}
-            itemCount={count}
-            loadMoreItems={loadMoreItems}
-            threshold={4}>
-            {({ onItemsRendered, ref }) => {
-              return (
-                <List
-                  height={height}
-                  itemCount={count}
-                  itemSize={56}
-                  onItemsRendered={onItemsRendered}
-                  ref={ref}
-                  width={width}>
-                  {EndpointRow}
-                </List>
-              );
-            }}
-          </InfiniteLoader>
-        );
-      }}
-    </AutoSizer>
+    <InfiniteScrollList
+      count={count}
+      loadMoreItems={loadMoreItems}
+      loaderRef={infiniteLoaderRef}
+      isItemLoaded={isItemLoaded}
+      listItems={endpoints}
+      handleListItemSelect={handleListItemSelect}
+      selectedEndpoint={selectedEndpoint}
+    />
   );
 };
