+++ apps/Conduit-UI/src/components/database/custom-endpoints/InfiniteScrollList.tsx
@@ -1,0 +1,144 @@
+import React, { FC } from 'react';
+import {
+  Box,
+  Typography,
+  Skeleton,
+  ListItemButton,
+  ListItemIcon,
+  Paper,
+  ListItemText,
+} from '@mui/material';
+import InfiniteLoader from 'react-window-infinite-loader';
+import { FixedSizeList as List, ListChildComponentProps } from 'react-window';
+import AutoSizer from 'react-virtualized-auto-sizer';
+import { OperationsEnum } from '../../../models/OperationsEnum';
+import { getOperation } from '../../../utils/getOperation';
+import { EndpointTypes, Schema } from '../../../models/database/CmsModels';
+
+interface Props {
+  listItems: EndpointTypes[] | Schema[];
+  count: number;
+  loaderRef: any;
+  isItemLoaded: (index: number) => boolean;
+  loadMoreItems: (startIndex: number, stopIndex: number) => void;
+  handleListItemSelect: (endpoint: string) => void;
+  selectedEndpoint?: Schema;
+  badge?: boolean;
+}
+
+const InfiniteScrollList: FC<Props> = ({
+  listItems,
+  count,
+  loaderRef,
+  isItemLoaded,
+  loadMoreItems,
+  handleListItemSelect,
+  selectedEndpoint,
+  badge,
+}) => {
+  const ListRow = ({ index, style }: ListChildComponentProps) => {
+    const listItem = listItems[index];
+
+    const getBadgeColor = (endpointForBadge: any) => {
+      switch (endpointForBadge.operation) {
+        case OperationsEnum.POST:
+          return '#49cc90';
+        case OperationsEnum.PUT:
+          return '#fca130';
+        case OperationsEnum.DELETE:
+          return '#f93e3e';
+        case OperationsEnum.GET:
+          return '#61affe';
+        case OperationsEnum.PATCH:
+          return '#50e3c2';
+      }
+    };
+
+    return (
+      <div style={style}>
+        {!listItem ? (
+          <Typography variant="h2">
+            <Skeleton />
+          </Typography>
+        ) : (
+          <ListItemButton
+            sx={{
+              '&.MuiListItemButton-root': {
+                '&.Mui-selected': {
+                  background: 'secondary',
+                },
+                borderRadius: '10px',
+              },
+            }}
+            key={`endpoint-${listItem._id}`}
+            onClick={() => handleListItemSelect(listItem.name)}
+            selected={selectedEndpoint?._id === listItem?._id}>
+            {badge ? (
+              <ListItemIcon>
+                <Paper
+                  elevation={12}
+                  sx={{
+                    color: 'white',
+                    padding: 0.2,
+                    width: '65px',
+                    textAlign: 'center',
+                    marginRight: 1,
+                    backgroundColor: getBadgeColor(listItem),
+                  }}>
+                  {getOperation(listItem)}
+                </Paper>
+              </ListItemIcon>
+            ) : null}
+
+            <ListItemText
+              primary={listItem.name}
+              primaryTypographyProps={{
+                style: { overflow: 'hidden', whiteSpace: 'nowrap', textOverflow: 'ellipsis' },
+              }}
+            />
+          </ListItemButton>
+        )}
+      </div>
+    );
+  };
+
+  return (
+    <AutoSizer>
+      {({ height, width }) => {
+        if (!count) {
+          return (
+            <Box width={width}>
+              <Typography sx={{ textAlign: 'center', marginTop: '100px' }}>
+                No endpoints available
+              </Typography>
+            </Box>
+          );
+        }
+        return (
+          <InfiniteLoader
+            ref={loaderRef}
+            isItemLoaded={isItemLoaded}
+            itemCount={count}
+            loadMoreItems={loadMoreItems}
+            threshold={4}>
+            {({ onItemsRendered, ref }) => {
+              return (
+                <List
+                  height={height}
+                  itemCount={count}
+                  itemSize={56}
+                  onItemsRendered={onItemsRendered}
+                  ref={ref}
+                  width={width}>
+                  {ListRow}
+                </List>
+              );
+            }}
+          </InfiniteLoader>
+        );
+      }}
+    </AutoSizer>
+  );
+};
+
+export default InfiniteScrollList;
